requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alpaca-py==0.43.1
anthropic==0.8.0
alpaca-trade-api==3.2.0
annotated-types==0.7.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
deprecation==2.1.0
frozenlist==1.8.0
httpx==0.25.2
idna==3.11
msgpack==1.0.3
multidict==6.7.0
numpy==2.3.4
openai==1.6.1
packaging==25.0
pandas==2.3.3
propcache==0.4.1
pydantic==2.12.3
pydantic-settings==2.1.0
pydantic_core==2.41.4
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.1
requests==2.32.5
six==1.17.0
sqlalchemy==2.0.36
sseclient-py==1.8.0
structlog==23.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
websocket-client==1.9.0
websockets==10.4
yarl==1.22.0
